requirements.txt
anthropic==0.39.0
feedparser==6.0.11
requests==2.32.3
PyYAML==6.0.2
python-dateutil==2.9.0.post0
Jinja2==3.1.4
beautifulsoup4==4.12.3
